+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#955](https://github.com/xmake-io/xmake/issues/955#issuecomment-766481512): Support `zig cc` and `zig c++` as c/c++ compiler
+* [#955](https://github.com/xmake-io/xmake/issues/955#issuecomment-768193083): Support zig cross-compilation
 * [#1177](https://github.com/xmake-io/xmake/issues/1177): Improve to detect terminal and color codes
 
 ## v2.5.1
@@ -910,6 +911,7 @@
 ### 新特性
 
 * [#955](https://github.com/xmake-io/xmake/issues/955#issuecomment-766481512): 支持 `zig cc` 和 `zig c++` 作为 c/c++ 编译器
+* [#955](https://github.com/xmake-io/xmake/issues/955#issuecomment-768193083): 支持使用 zig 进行交叉编译
 * [#1177](https://github.com/xmake-io/xmake/issues/1177): 改进终端和 color codes 探测
 
 ## v2.5.1